+++ History.md
@@ -5,6 +5,7 @@ This incorporates all changes after 4.13.1 up to 4.14.0.
 
   * remove:
     - `res.json(status, obj)` signature - use `res.status(status).json(obj)`
+    - `res.jsonp(status, obj)` signature - use `res.status(status).jsonp(obj)`
     - `res.vary()` (no arguments) -- provide a field name as an argument
 
 5.0.0-alpha.2 / 2015-07-06